feat: complete docker + initializing commant

+++ src/catalog/catalog.controller.ts
@@ -3,55 +3,55 @@ import { CatalogService } from './catalog.service';
 import { CreateCatalogDto } from './dto/create-catalog.dto';
 import { UpdateCatalogDto } from './dto/update-catalog.dto';
 import { ApiBearerAuth, ApiCreatedResponse, ApiOkResponse } from '@nestjs/swagger';
-import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 import { CatalogEntity } from './entities/catalog.entity';
 import { AuthorCatalog } from './dto/author-catalog.dto';
+import { BookCatalog } from './dto/book-catalog.dto';
+import { Roles } from 'src/auth/guards/role/roles.decorator';
+import { Role } from 'src/auth/guards/role/role.enum';
 
 @Controller('catalog')
 export class CatalogController {
   constructor(private catalogService: CatalogService) {}
 
   @Post()
-  @UseGuards(JwtAuthGuard)
+  @Roles(Role.Admin)
   @ApiCreatedResponse({ type: CatalogEntity })
   @ApiBearerAuth()
   async create(@Body() createCatalogDto: CreateCatalogDto) {
-    // return new CatalogEntity(await this.catalogService.create(createCatalogDto));
     return await this.catalogService.create(createCatalogDto)
   }
 
-  @Get()
-  @UseGuards(JwtAuthGuard)
+  @Post('AuthorsBooks')
+  @Roles(Role.User)
+  @ApiCreatedResponse({ type: CatalogEntity })
   @ApiOkResponse({ type: CatalogEntity, isArray: true })
+  @ApiBearerAuth()
   async getBooksByAuthor(@Body() authorCatalog: AuthorCatalog) {
-    return await this.catalogService.getBookByAuthor(authorCatalog)
+    return await this.catalogService.getBooksByAuthor(authorCatalog)
   }
 
-  // TO DO !@#@!#!@
-  // @Get()
-  // @UseGuards(JwtAuthGuard)
-  // @ApiOkResponse({ type: CatalogEntity, isArray: true })
-  // async getBookByNaming() {
-  //   const all = await this.catalogService.findAll();
-  //   return all.map((elem) => new CatalogEntity(elem))
-  // }
+  @Post('booksByNaming')
+  @Roles(Role.User)
+  @ApiCreatedResponse({ type: CatalogEntity })
+  @ApiOkResponse({ type: CatalogEntity, isArray: true })
+  @ApiBearerAuth()
+  async getBooksByNaming(@Body() bookCatalog: BookCatalog) {
+    return await this.catalogService.getBooksByNaming(bookCatalog)
+  }
 
   @Get()
-  @UseGuards(JwtAuthGuard)
+  @Roles(Role.Admin)
   @ApiOkResponse({ type: CatalogEntity, isArray: true })
+  @ApiBearerAuth()
   async findAll() {
-    // const all = await this.catalogService.findAll();
-    // return all.map((elem) => new CatalogEntity(elem))
-
     return await this.catalogService.findAll();
   }
 
   @Get(':id')
-  // @UseGuards(JwtAuthGuard)
+  @Roles(Role.Admin)
   @ApiOkResponse({ type: CatalogEntity })
+  @ApiBearerAuth()
   async findOne(@Param('id', ParseIntPipe) id: number) {
-
-
     const catalog = await this.catalogService.findOne(id)
 
     if(!catalog) {
@@ -59,31 +59,37 @@ export class CatalogController {
     }
 
     return catalog
-
-    // return new CatalogEntity(article)
   }
 
   @Patch(':id')
-  @UseGuards(JwtAuthGuard)
+  @Roles(Role.Admin)
   @ApiBearerAuth()
   @ApiCreatedResponse({ type: CatalogEntity })
   async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateCatalogDto: UpdateCatalogDto,
   ) {
+    const catalog = await this.catalogService.findOne(id)
+
+    if(!catalog) {
+      throw new NotFoundException(`catalog with ${id} does not exist.`)
+    }
+
     return await this.catalogService.update(id, updateCatalogDto)
-    // 
-    // return new CatalogEntity(
-    //   await this.catalogService.update(id, updateArticleDto),
-    // );
   }
 
   @Delete(':id')
-  @UseGuards(JwtAuthGuard)
+  @Roles(Role.Admin)
   @ApiBearerAuth()
   @ApiOkResponse({ type: CatalogEntity })
   async remove(@Param('id', ParseIntPipe) id: number) {
+
+    const catalog = await this.catalogService.findOne(id)
+
+    if(!catalog) {
+      throw new NotFoundException(`catalog with ${id} does not exist.`)
+    }
+
     return await this.catalogService.remove(id)
-    // return new CatalogEntity(await this.catalogService.remove(id));
   }
 }
